strlen修正

+++ c++/src/string_ex.c
@@ -15,7 +15,7 @@ void lntrim(char *str) {
 int searchc(char *str, char c){
 	int i, cnt = 0;
 
-	for(i=0; i<strlen(str); i++){
+	for(i=0; i<(int)strlen(str); i++){
 		if(str[i] == c) cnt++;
 	}
 	return cnt;
@@ -41,7 +41,7 @@ int split(char *str, char c, char *outlist[] ) {
 int delch(char *str, char c, char *out) {
 	int i=0, j=0;
 
-	for (i=0; i<strlen(str); i++) {
+	for (i=0; i<(int)strlen(str); i++) {
 		if (str[i] == c) {
 			j++;
 			continue;
